Changing VIM back to dim in MEAN SHEAR
Currently the loops in the SHEAR post-processing calculation have a "dim" - it seems to me this should be VIM especially for SWIRLING calculations.  But that change created velotheta exodus II comparison failure.  So I changed it back temporarily.  Not sure why a VIM would change velotheta's results (just a cylindrical 2D problem)