Add log hunting to input, various cleanup (#211)
* Gomathon_June2018

* June2018B

* June2018C

* 23Oct2018

* Residence Time field through INTD and Radiative Transfer Eqn fixes.

* dp_vif for VELO_SPEED

* Thanksgiving

* Dec10-2018

* Changing VIM back to dim in MEAN SHEAR

Currently the loops in the SHEAR post-processing calculation have a "dim" - it seems to me this should be VIM especially for SWIRLING calculations.  But that change created velotheta exodus II comparison failure.  So I changed it back temporarily.  Not sure why a VIM would change velotheta's results (just a cylindrical 2D problem)

* Update mm_post_proc.c

* Dec102018

* updating mm_post_proc.c for 2nd Invariant