Gas egress (#494)
Adding "gas egress" channels as a means to promote wetting (lower dynamic contact angles) with multi-phase lubrication calculations where sidewall drag is included. On the liquid side of the LS interface, the lubrication gap is reduced to capture the increased drag of the sidewall. On the gas side, the opposite is true. The gap near the side wall is increased to promote gas egress from the interface zone and thereby lower the apparent dynamic contact angle.